+++ src/model/markdown-file.ts
@@ -21,6 +21,7 @@ export class MarkdownFile extends OutputFile {
   public addText(text: string) {
     //super.addText(this._replaceInlineReferences(text), encodeHtml);
     super.addText(text);
+    return this;
   }
 
   public addLink(text: string) {
